requirements.txt
transformers>=4.34.0
sentence-transformers>=2.2.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
qdrant-client>=1.6.0
PyPDF2>=3.0.0
pymupdf>=1.23.0
python-docx>=0.8.11
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.28.0
tqdm>=4.65.0
python-dotenv>=1.0.0
psutil>=5.9.0
sentencepiece>=0.1.99